+++ ios-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.98](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.97...@times-components/ios-app@0.17.98) (2020-05-05)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 ## [0.17.97](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.95...@times-components/ios-app@0.17.97) (2020-04-30)