+++ packages/article/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.4.91](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.90...@times-components/article@7.4.91) (2020-05-05)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.4.90](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.89...@times-components/article@7.4.90) (2020-04-30)
 
 **Note:** Version bump only for package @times-components/article